fix: change double to bigdecial coupon dto

+++ src/main/java/com/onebrain/onebrain/dto/CouponRequest.java
@@ -4,6 +4,8 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
 
+import java.math.BigDecimal;
+
 public record CouponRequest(
         @NotNull()
         @Size(min = 6)
@@ -13,7 +15,7 @@ public record CouponRequest(
         String description,
 
         @NotNull
-        Double discountValue,
+        BigDecimal discountValue,
 
         @NotNull
         String expirationDate,
